Admin policies

+++ app/Http/Controllers/AdminController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
 
 use App\Models\User;
 use App\Models\Tag;
@@ -13,7 +14,7 @@ use App\Models\Admin;
 class AdminController extends Controller
 {
     public function showDashboard(){
-        //policy
+        $this->authorize('admin', Admin::class);
 
         $users = User::all();
         $tags = Tag::all();
@@ -30,13 +31,15 @@ class AdminController extends Controller
         return view('pages.manageusers', ['users' => $users]);
     }
 
+
     public function showCreateUser(){
-        //policy
+        $this->authorize('admin', Admin::class);
 
         return view('pages.createUser');
     }
 
     public function createUser(Request $request){
+        $this->authorize('admin', Admin::class);
 
         $request->validate([
             'name' => 'required|string|max:255',
@@ -56,10 +59,9 @@ class AdminController extends Controller
     }
 
     public function search(Request $request){
-        //policy
 
         $input =$request->input('user_search_query');
-        $users = User::select('users.name', 'users.user_id', 'users.email', 'users.moderator')
+        $users = User::select('users.name', 'users.user_id', 'users.email', 'users.is_moderator')
                     ->where('name', 'ilike', '%' . $input . '%')
                     ->get();
 
@@ -67,9 +69,9 @@ class AdminController extends Controller
     }
 
     public function deleteUser(int $id){
-        $user = User::find($id);
+        $this->authorize('admin', Admin::class);
 
-        //policy
+        $user = User::find($id);
 
         $user->delete();
 
@@ -79,7 +81,7 @@ class AdminController extends Controller
     public function blockUser(int $id){
         $user = User::find($id);
 
-        //policy
+        $this->authorize('admin', Admin::class);
 
         $user->is_blocked = true;
         $user->save();
@@ -90,7 +92,7 @@ class AdminController extends Controller
     public function unblockUser(int $id){
         $user = User::find($id);
 
-        //policy
+        $this->authorize('admin', Admin::class);
 
         $user->is_blocked = false;
         $user->save();
@@ -101,7 +103,7 @@ class AdminController extends Controller
     public function changeRole(int $id, Request $request){
         $user = User::find($id);
 
-        //policy
+        $this->authorize('admin', Admin::class);
 
         if($request->role === 'user'){
             $user->is_moderator = false;
